+++ scripts/articles/articleCreator.js
@@ -3,7 +3,7 @@
 
 //imports
 const articleFactory = require("./articleFactory")
-const Database = require("./Database")
+const Database = require("../Database")
 
 //Create empty array if we need it
 Database.articles = Database.articles || []
