+++ scripts/main.js
@@ -1,6 +1,6 @@
 // Authors : Chase, John, Max, and Paul
 // Purpose : ¯\_(ツ)_/¯
 const Database = require("./Database");
-const usersRegister = require("./usersRegister");
-const usersLogin = require("./usersLogin");
+const usersRegister = require("./users/usersRegister");
+const usersLogin = require("./users/usersLogin");
