+++ scripts/users/usersFactory.js
@@ -1,6 +1,6 @@
 //Author: Paul Ellis; Purpose: function that creates new users
-const Database = require("./Database");
-const idGenerator = require("./idGenerator");
+const Database = require("../Database");
+const idGenerator = require("../idGenerator");
 const userIdGen = idGenerator();
 
 const usersFactory = (name, email) => {
